@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
@@ -29,6 +27,12 @@ from .user import BaseUser
 from .activity import create_activity
 from .invite import Invite
 from .enums import Status, try_enum
+
+__all__ = (
+    'WidgetChannel',
+    'WidgetMember',
+    'Widget',
+)
 
 class WidgetChannel:
     """Represents a "partial" widget channel.
@@ -77,7 +81,7 @@ class WidgetChannel:
     @property
     def mention(self):
         """:class:`str`: The string that allows you to mention the channel."""
-        return '<#%s>' % self.id
+        return f'<#{self.id}>'
 
     @property
     def created_at(self):
@@ -236,7 +240,7 @@ class Widget:
     @property
     def json_url(self):
         """:class:`str`: The JSON URL of the widget."""
-        return "https://discord.com/api/guilds/{0.id}/widget.json".format(self)
+        return f"https://discord.com/api/guilds/{self.id}/widget.json"
 
     @property
     def invite_url(self):
