@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
@@ -26,6 +24,10 @@ DEALINGS IN THE SOFTWARE.
 
 import time
 import random
+
+__all__ = (
+    'ExponentialBackoff',
+)
 
 class ExponentialBackoff:
     """An implementation of the exponential backoff algorithm
