@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 discord.ext.commands
 ~~~~~~~~~~~~~~~~~~~~~
@@ -10,8 +8,8 @@ An extension module to facilitate creation of bot commands.
 :license: MIT, see LICENSE for more details.
 """
 
-from .bot import Bot, AutoShardedBot, when_mentioned, when_mentioned_or
-from .context import Context
+from .bot import *
+from .context import *
 from .core import *
 from .errors import *
 from .help import *
