@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
@@ -25,6 +23,10 @@ DEALINGS IN THE SOFTWARE.
 """
 
 from .iterators import ReactionIterator
+
+__all__ = (
+    'Reaction',
+)
 
 class Reaction:
     """Represents a reaction to a message.
