@@ -1,5 +1,3 @@
-# -*- coding: utf-8 -*-
-
 """
 The MIT License (MIT)
 
@@ -28,6 +26,10 @@ from .mixins import Hashable
 from .asset import Asset
 from .utils import snowflake_time
 from .enums import StickerType, try_enum
+
+__all__ = (
+    'Sticker',
+)
 
 class Sticker(Hashable):
     """Represents a sticker.
@@ -93,7 +95,7 @@ class Sticker(Hashable):
 
     @property
     def created_at(self):
-        """:class:`datetime.datetime`: Returns the sticker's creation time in UTC as a naive datetime."""
+        """:class:`datetime.datetime`: Returns the sticker's creation time in UTC."""
         return snowflake_time(self.id)
 
     @property
